Improve names.
Existing name of a device takes the room name and appends the
device name. For multi-channel devices this ends up with more
than one light with the same name. To fix the channel number is also
appended to the name for multi-channel devices.

Additionally, the parsing of the UPE file has a new feature. The file
may be manually edited adding new records to the end of the file.
The records can be used to rename lights if you don't like the name
that is generated.

The record format is:
  99,<device id>,<name>

For example if you wish to rename device 142_42_0 to
"Funky new name" you could create a record in the UPE file as so:
99,142_42_0,Funky new name